Save as SVG, TXT, and remove generation of JSON file

+++ run.py
@@ -1,3 +1,4 @@
+import json
 import os
 import re
 from flask import Flask, render_template, request
@@ -40,25 +41,31 @@ def index():
         print(f'fm_file: {fm_file}')
 
         if not fm_file:
+            # The file is required and this is controlled in the front-end.
             pass
 
-        if fm_file:
-            filename = fm_file.filename
-            try:
-                fm_file.save(filename)
-                
-                fm = FeatureIDEReader(filename).transform() 
-                if not name:
-                    name = os.path.splitext(os.path.basename(filename))[0]
-                
-                fm_characterization = FMCharacterization(fm, name)
-                fm_characterization.set_metadata(name, description, author, reference, keywords)
-                interfaces.to_json(fm_characterization, FM_FACT_JSON_FILE)
-            except:
-                data['fm_facts'] = False
-                os.remove(filename)
+        filename = fm_file.filename
+        try:
+            fm_file.save(filename)
+            
+            fm = FeatureIDEReader(filename).transform() 
+            if not name:
+                name = os.path.splitext(os.path.basename(filename))[0]
+            
+            characterization = FMCharacterization(fm, name)
+            characterization.set_metadata(name, description, author, reference, keywords)
+            #json_characterization = interfaces.to_json(fm_characterization, FM_FACT_JSON_FILE)
+            json_characterization = interfaces.to_json(characterization)
+            str_characterization = interfaces.get_string_output(characterization)
+            data['fm_facts'] = json_characterization
+            data['fm_characterization'] = str_characterization
+        except Exception as e:
+            data = None
+            print(e)
 
-        data['fm_facts'] = True
+        if os.path.exists(filename):
+            os.remove(filename)
+        
         return render_template('index.html', data=data)
